Initialize a Vite React app, no TS, in web-js
npm create vite@latest web-js -- --template react

TypeScript is turning into too much of a time sink, for too little
benefit. I'll accept the risk that an argument might be null in
production unexpectedly, in exchange for being able to do something
other than fight with TypeScript for as much time as I am fighting with
it.