Expect a login button
There are too many useless words. Gramming is going to involve combing
through the dictionary and picking which words you want to use and not
use. This is more than is likely to be done in one browser session. So
people will need to log in to use the app.