Remove the this.account output
It seems like this.account cannot be trusted unless you're using the
value directly.

If I do an output of

value: this.account.slice(0,20)

I get

Value: ${Token[AWS.AccountI

With

value: this.account

I get

Value:
  Ref: AWS::AccountId

I guess I'm getting some kind of token or even memory location when I
try to use this.account in calculations, and this can vary depending on
how exactly I use it.

So it's better to not feed the account id into the computation of the
domain name prefix.

It's good enough that I'm varying it with the repo and branch.