Add EmailIdentities
gh-20

I have the domain names set up for test and prod. These docs:

https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_ses-readme.html#email-identity

specifically the part where it says

> To verify an identity for a hosted zone, you create an EmailIdentity:

make it sound like maybe AWS/CDK will kick off the verification process
for you, like it does with TLS certificates. That would be great.

Even if it doesn't, I'd prefer to create the EmailIdentities that I'll
use for the from-address for sign-up verification in CDK, as I'm doing
here.